Canvas support for paths & drawing lines and circles to paths.
Now you can draw a line or a circle to a path (a series of points)
and reuse those points to redraw the same figure multiple times in
different places. You can use one path as a pattern to draw other
paths, for instance, draw circles along the path of a straight line.
Or a circle with circles all around!